Update forum thread link
A change of the software used for the Arduino Forum caused the format of the topic URLs to change. Even though the old
ones still redirect, it is safest to not rely on the persistence of such redirects. So the URL of the reference forum
topic is hereby updated to the new format.