Document project status
My testing indicates that the workaround is no longer needed with modern versions of AVRDUDE. This fact, combined with
the incompatibility of the project with modern versions of the Arduino development software means there is no point in
maintaining the project.

The status is hereby clearly documented in the readme.